refactor api gateway code for response reuse, update documentation
